Fix wrap for AUDIO_INT
Floats don't have enough precision for all 32 bit integers. In
particular, near INT32_MAX their value is INT32_MAX + 1, which, when
converted back to int becomes INT32_MIN.

+++ src/add-ons/media/media-add-ons/mixer/Resampler.cpp
@@ -33,7 +33,7 @@ kernel(Resampler* object, const void *_src, int32 srcSampleOffset,
 	if (srcSampleCount == destSampleCount) {
 		// optimized case for no resampling
 		while (count--) {
-			float tmp = ((*(const inType*)src) - inMiddle) * gain + outMiddle;
+			double tmp = ((*(const inType*)src) - inMiddle) * gain + outMiddle;
 			if (tmp < min) tmp = min;
 			if (tmp > max) tmp = max;
 			*(outType *)dest = (outType)tmp;
@@ -48,7 +48,7 @@ kernel(Resampler* object, const void *_src, int32 srcSampleOffset,
 
 	// downsample
 	while (count--) {
-		float tmp = ((*(const inType*)src) - inMiddle) * gain + outMiddle;
+		double tmp = ((*(const inType*)src) - inMiddle) * gain + outMiddle;
 		if (tmp < min) tmp = min;
 		if (tmp > max) tmp = max;
 		*(outType *)dest = (outType)tmp;
